+++ src/api/api.ts
@@ -82,7 +82,7 @@ export class MastoApi {
             this.fetchHomeFeed(numTimelineToots, maxId),
         ];
 
-        // Only retrieve trending toots on the first call to this method
+        // Only fetch trending toots first time this is called (skip when paging through timeline)
         if (!maxId) {
             promises = promises.concat([
                 fetchRecentTootsForTrendingTags(),
@@ -92,7 +92,7 @@ export class MastoApi {
 
         const allResponses = await Promise.all(promises);
         console.debug(`[MastoApi] getFeed() allResponses: ${JSON.stringify(allResponses, null, 4)}`);
-        const homeToots = allResponses.shift();
+        const homeToots = allResponses.shift();  // Pop timeline toots off the array
 
         return {
             homeToots: homeToots,